+++ src/ETHTelnet.cpp
@@ -1,2 +1,2 @@
 #define TELNET_USE_ETHERNET
-#include "ESPTelnet.cpp"
+#include "Telnet.cpp"